fixing incorrect package name [io.appform.ranger.zk.server.bundle] -> [io.appform.ranger.server.bundle]
