+++ ranger-server-bundle/src/main/java/io/appform/ranger/server/bundle/rotation/OorTask.java
@@ -13,27 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.appform.ranger.zk.server.bundle.healthcheck;
+package io.appform.ranger.server.bundle.rotation;
 
-import com.codahale.metrics.health.HealthCheck;
+import io.dropwizard.servlets.tasks.Task;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.curator.framework.CuratorFramework;
 
-import javax.inject.Singleton;
+import java.io.PrintWriter;
+import java.util.List;
+import java.util.Map;
 
-@Singleton
 @Slf4j
-public class RangerHealthCheck extends HealthCheck {
+public class OorTask extends Task {
 
-    private final CuratorFramework curatorFramework;
+    private final RotationStatus rotationStatus;
 
-    public RangerHealthCheck(CuratorFramework curatorFramework){
-        this.curatorFramework = curatorFramework;
+    public OorTask(RotationStatus rotationStatus) {
+        super("ranger-oor");
+        this.rotationStatus = rotationStatus;
     }
 
     @Override
-    protected Result check() {
-        return curatorFramework.getZookeeperClient().isConnected() ?
-                Result.healthy("Service is healthy") : Result.unhealthy("Can't connect to zookeeper");
+    public void execute(Map<String, List<String>> map, PrintWriter printWriter) {
+        rotationStatus.oor();
+        log.info("Taking node out of rotation on ranger");
     }
 }
