+++ ranger-server-bundle/src/main/java/io/appform/ranger/server/bundle/rotation/RotationStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.appform.ranger.zk.server.bundle.rotation;
+package io.appform.ranger.server.bundle.rotation;
 
 import java.util.concurrent.atomic.AtomicBoolean;
