+++ ranger-zk-server-bundle/src/main/java/io/appform/ranger/server/bundle/lifecycle/CuratorLifecycle.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.appform.ranger.zk.server.bundle.lifecycle;
+package io.appform.ranger.server.bundle.lifecycle;
 
 import io.appform.ranger.common.server.ShardInfo;
 import io.appform.ranger.core.signals.Signal;
