+++ ranger-zk-server/src/main/java/io/appform/ranger/zk/server/App.java
@@ -15,8 +15,8 @@
  */
 package io.appform.ranger.zk.server;
 
-import io.appform.ranger.zk.server.bundle.ZKServerBundle;
-import io.appform.ranger.zk.server.bundle.config.RangerConfiguration;
+import io.appform.ranger.server.bundle.ZKServerBundle;
+import io.appform.ranger.server.bundle.config.RangerConfiguration;
 import io.dropwizard.Application;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
